plotting quality assessment per radius

Plotting the time series and cross-correlation to contrast each radius with the chronologies, using both raw ring-width measurements and treated series.

plot_qa(qa.trt, qa.out_series = "all")

Arguments

qa.trt

object "cfs_qa" from CFS_qa() function.

qa.out_series

series_id list to be plotted, default "all" for plotting the graphs for all.

Value

A named list of tree-ring series, where each element corresponds to a series requested via qa.out_series. Each element is itself a list containing four ggplot objects:

plot.raw.series

Raw tree-ring series vs year.

plot.trt.series

Treated (detrended/standardized) series vs year.

plot.raw.ccf

Cross-correlation function of the raw series.

plot.trt.ccf

Cross-correlation function of the treated series.

Examples

# \donttest{

# loading processed data
dt.samples_trt <- readRDS(system.file("extdata", "dt.samples_trt.rds", package = "growthTrendR"))
# data processing
dt.samples_long <- prepare_samples_clim(
dt.samples_trt = dt.samples_trt, calbai = FALSE )
# rename to the reserved column name
data.table::setnames(
dt.samples_long,
c("sample_id", "year", "rw_mm"),
c("SampleID", "Year" ,"RawRing"))
# assign treated series
# users can decide their own treated series

# for rhub::rhub_check() on macos VECTOR_ELT issues
data.table::setorder(dt.samples_long, SampleID, Year)
dt.samples_long$RW_trt <-
  ave(
  as.numeric(dt.samples_long$RawRing),
  dt.samples_long$SampleID,
  FUN = function(x)
  if (length(x) > 1L) c(NA_real_, diff(x)) else NA_real_
  )
# quality check on radius alignment based on the treated series
dt.qa <-CFS_qa(dt.input = dt.samples_long, qa.label_data = "demo-samples",
qa.label_trt = "difference", qa.min_nseries = 5)
#> ============================================================
#> STEP 1: Computing pair-wise CCF with auto-batching
#> ============================================================
#> Progress pair-wise ccf...
#> Step 1 complete: 9 out of 9 initial candidate samples identified
#> 
#> ============================================================
#> STEP 2: Iterative refinement of chronologies
#> ============================================================
#> Step 2 complete: Converged in 1 iterations (success)
#> 
#> ============================================================
#> CFS_qa analysis complete!
#> ============================================================
plots.lst <- plot_qa(dt.qa, qa.out_series = "X003_101_005")
# }